// frontend-tsx/src/_state/actions-types/rooms/types.tsx
export enum RoomActionType {
    CREATE_ROOM_SUCCESS = "CREATE_ROOM_SUCCESS",
    CREATE_ROOM_FAIL = "CREATE_ROOM_FAIL",
    UPDATE_ROOM_SUCCESS = "UPDATE_ROOM_SUCCESS",
    UPDATE_ROOM_FAIL = "UPDATE_ROOM_FAIL",
    JOIN_ROOM_SUCCESS = "JOIN_ROOM_SUCCESS",
    JOIN_ROOM_FAIL = "JOIN_ROOM_FAIL",
    LOAD_ROOM_DETAILS_SUCCESS = "LOAD_ROOM_DETAILS_SUCCESS",
    LOAD_ROOM_DETAILS_FAIL = "LOAD_ROOM_DETAILS_FAIL",
}